Implement first version (non-functional) of gRPC producer
