+++ pulsar-protocols/grpc/src/main/java/org/apache/pulsar/protocols/grpc/GrpcServer.java
@@ -5,19 +5,21 @@ import io.grpc.ManagedChannelBuilder;
 import io.grpc.Server;
 import io.grpc.ServerBuilder;
 import io.grpc.protobuf.services.ProtoReflectionService;
-import lombok.extern.slf4j.Slf4j;
 import org.apache.pulsar.broker.service.BrokerService;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import java.io.IOException;
 import java.util.concurrent.TimeUnit;
 
-@Slf4j
 public class GrpcServer {
+
+    private static final Logger log = LoggerFactory.getLogger(GrpcServer.class);
 
     private Server server;
 
     public void start(BrokerService service) throws IOException {
-        /* The port on which the server should run */
+        // TODO: replace with configurable port
         int port = 50444;
         server = ServerBuilder.forPort(port)
                 .addService(new PulsarGrpcService(service))
@@ -25,19 +27,16 @@ public class GrpcServer {
                 .build()
                 .start();
         log.info("############# Server started, listening on " + port);
-        Runtime.getRuntime().addShutdownHook(new Thread() {
-            @Override
-            public void run() {
-                // Use stderr here since the logger may have been reset by its JVM shutdown hook.
-                log.error("############### shutting down gRPC server since JVM is shutting down");
-                try {
-                    GrpcServer.this.stop();
-                } catch (InterruptedException e) {
-                    e.printStackTrace(System.err);
-                }
-                log.error("############### gRPC server shut down");
+        Runtime.getRuntime().addShutdownHook(new Thread(() -> {
+            // Use stderr here since the logger may have been reset by its JVM shutdown hook.
+            log.error("############### shutting down gRPC server since JVM is shutting down");
+            try {
+                GrpcServer.this.stop();
+            } catch (InterruptedException e) {
+                e.printStackTrace(System.err);
             }
-        });
+            log.error("############### gRPC server shut down");
+        }));
     }
 
     public void stop() throws InterruptedException {
