+++ pulsar-protocols/grpc/src/main/proto/PulsarGrpcService.proto
@@ -2,12 +2,15 @@ syntax = "proto3";
 
 option java_multiple_files = true;
 option java_package = "org.apache.pulsar.protocols.grpc";
-option optimize_for = SPEED;
+// option optimize_for = LITE_RUNTIME;
 
 package pulsar.protocols.grpc;
+import 'PulsarApi.proto';
 
 service PulsarGrpcService {
     rpc Hello(SimpleValue) returns (SimpleValue) {}
+    rpc produce(stream pulsar.proto.BaseCommand) returns (stream pulsar.proto.BaseCommand) {}
+    rpc consume(stream pulsar.proto.BaseCommand) returns (stream pulsar.proto.BaseCommand) {}
 }
 
 message SimpleValue {
